+++ CHANGELOG.md
@@ -14,7 +14,9 @@ These changes likely break some apps, please read the
   * Size and Position units were changed from int to float32
   * `Text.TextSize` moved to float32 and `fyne.MeasureText` now takes a float32 size parameter
   * Removed `Size.Union`
-  * Added fyne.Vector for generic X, Y storage
+  * Added fyne.Delta for difference based X, Y representation
+  * DraggedEvent.DraggedX and DraggedX (int, int) to DraggedEvent.Dragged (Delta)
+  * ScrollEvent.DeltaX and DeltaY (int, int) moved to ScrollEvent.Scrolled (Delta)
 
 * Theme API update
   * `fyne.Theme` moved to `fyne.LegacyTheme` and can be load to a new theme using `theme.FromLegacy`
@@ -31,7 +33,7 @@ These changes likely break some apps, please read the
 ### Changed
 
 * Coordinate system is now float32 - see breaking changes above
-* ScrollEvent and DragEvent moved to Vector from (int, int)
+* ScrollEvent and DragEvent moved to Delta from (int, int)
 
 * Change bundled resources to use more efficient string storage
 * Desktop left and right mouse buttons renamed to MouseButtonPrimary and MouseButtonSecondary